requirements.txt
transformers==4.40.0
torch==2.2.0
paramiko==3.4.0
python-nmap==0.7.1
python-telegram-bot==20.8
psutil==5.9.8
PyAutoGUI==0.9.54
twilio==8.0.0
playwright==1.40.0
aioshortcuts==0.1.5
paho-mqtt==1.6.1
fastapi==0.109.2
uvicorn==0.27.0
